Add files via upload
localhost/crosswalk displays Jedi Michael on the browser as it should. TWalk is in the browser head. The phpmyadmin has user table setup but my server xampp which seems to be working fine is not connecting with the database. I have tried everything I can think of to fix this.  I should be able to do a registration then login and it then show Jedi Michael which is the index page. Can someone help me to understand what the problem is?